+++ tests/base.py
@@ -42,7 +42,7 @@ from wxc_sdk.workspaces import Workspace, CallingType
 log = logging.getLogger(__name__)
 
 __all__ = ['TestCaseWithTokens', 'TestCaseWithLog', 'gather', 'TestWithLocations', 'TestCaseWithUsers', 'get_tokens',
-           'async_test', 'LoggedRequest', 'TestCaseWithUsersAndSpaces']
+           'async_test', 'LoggedRequest', 'TestCaseWithUsersAndSpaces', 'TestLocationsUsersWorkspacesVirtualLines']
 
 
 def gather(mapping: Iterable[Any], return_exceptions: bool = False) -> Generator[Union[Any, Exception]]: