+++ wxc_sdk/person_settings/common.py
@@ -81,9 +81,10 @@ class PersonSettingsApiChild(ApiChild, base=''):
                       ('people', 'outgoingPermission/digitPatterns'): ('telephony/config/people', '/'),
                       ('people', 'callBridge'): ('telephony/config/people', '/features/'),
                       ('people', 'agent'): ('telephony/config/people', '/'),
+                      ('people', 'musicOnHold'): ('telephony/config/people', '/'),
                       }
-        selector, feature_prefix = alternates.get((selector, self.feature), (selector, feature_prefix))
         if selector == 'people' and self.feature == 'voicemail' and path == 'passcode':
             # this is a new endpoint for users and is the only VM endpoint with a different URL structure
             return self.session.ep(f'telephony/config/people/{person_id}/voicemail/passcode')
+        selector, feature_prefix = alternates.get((selector, self.feature), (selector, feature_prefix))
         return self.session.ep(f'{selector}/{person_id}{feature_prefix}{self.feature}{path}')